# packages/backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /organizations/{organizationId}/tenants:
    post:
      operationId: OrganizationsAdminController_createTenant
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantDto'
      tags: &ref_21
        - organizations
      security: &ref_22
        - accessToken: []
  /organizations/{organizationId}/portfolios:
    post:
      operationId: OrganizationsAdminController_createPortfolio
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortfolioDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioDto'
      tags: *ref_21
      security: *ref_22
  /organizations/{organizationId}/properties:
    post:
      operationId: OrganizationsAdminController_createProperty
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePropertyDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyDto'
      tags: *ref_21
      security: *ref_22
  /organizations/{organizationId}/units:
    post:
      operationId: OrganizationsAdminController_createUnit
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUnitDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartialUnitDto'
      tags: *ref_21
      security: *ref_22
  /organizations/{organizationId}/leases:
    post:
      operationId: OrganizationsAdminController_createLease
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeaseDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartialLeaseDto'
      tags: *ref_21
      security: *ref_22
  /organizations/{organizationId}/leaseInvoices:
    post:
      operationId: OrganizationsAdminController_createLeaseInvoice
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeaseInvoiceDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartialLeaseInvoiceDto'
      tags: *ref_21
      security: *ref_22
  /organizations/{organizationId}/lease/{id}/leaseInvoices:
    post:
      operationId: OrganizationsAdminController_createInvoices
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateManyLeaseInvoicesDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: string
      tags: *ref_21
      security: *ref_22
  /organizations/{organizationId}/expenses:
    post:
      operationId: OrganizationsAdminController_createExpense
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartialExpenseDto'
      tags: *ref_21
      security: *ref_22
  /organizations/{organizationId}/payouts:
    post:
      operationId: OrganizationsAdminController_createPayout
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePayoutDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutDto'
      tags: *ref_21
      security: *ref_22
  /organizations/{organizationId}/leaseInvoices/{id}/send-invoice-email:
    post:
      operationId: OrganizationsAdminController_sendInvoiceEmail
      summary: ''
      description: ''
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      tags: *ref_21
      security: *ref_22
  /organizations/{organizationId}/files:
    post:
      operationId: FilesController_create
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateFileDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: string
      tags: &ref_23
        - files
      security: &ref_24
        - accessToken: []
    get:
      operationId: FilesController_findAll
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
        - name: relationKey
          required: true
          in: query
          schema:
            $ref: '#/components/schemas/FileRelationKeyEnum'
        - name: relationValue
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                title: PaginatedFileDto
                required:
                  - results
                allOf:
                  - $ref: '#/components/schemas/PaginatedMetaDto'
                  - title: ArrayOfFileDto
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/FileDto'
      tags: *ref_23
      security: *ref_24
    delete:
      operationId: FilesController_remove
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
        - name: id
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
      tags: *ref_23
      security: *ref_24
  /organizations/{organizationId}/files/find-one:
    get:
      operationId: FilesController_findOne
      summary: ''
      description: ''
      parameters:
        - name: organizationId
          required: true
          in: path
          schema:
            type: string
        - name: id
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
      tags: *ref_23
      security: *ref_24
...
info:
  title: Aqaratech
  description: Aqaratech API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: accessToken
  schemas:
    TenantDto:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        fullName:
          type: string
        label:
          type: string
          nullable: true
        civilid:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        dob:
          format: date-time
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        passportNum:
          type: string
          nullable: true
        residencyEnd:
          format: date-time
          type: string
          nullable: true
        residencyNum:
          type: string
          nullable: true
        organizationId:
          type: string
        title:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - fullName
        - label
        - civilid
        - phone
        - dob
        - nationality
        - passportNum
        - residencyEnd
        - residencyNum
        - organizationId
        - title
        - createdAt
        - updatedAt
...
    PartialLeaseInvoiceDto:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        breadcrumbs:
          $ref: '#/components/schemas/LeaseInvoiceBreadcrumbsDto'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        organizationId:
          type: string
        portfolioId:
          type: string
        amount:
          type: number
        postAt:
          format: date-time
          type: string
        leaseId:
          type: string
        dueAt:
          format: date-time
          type: string
          nullable: true
        paidAt:
          format: date-time
          type: string
          nullable: true
        isPaid:
          type: boolean
        memo:
          type: string
          nullable: true
        mfPaymentId:
          type: string
          nullable: true
...
    PartialLeaseDto:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        breadcrumbs:
          $ref: '#/components/schemas/LeaseBreadcrumbsDto'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        organizationId:
          type: string
        portfolioId:
          type: string
        tenantId:
          type: string
        unitId:
          type: string
        start:
          format: date-time
          type: string
        end:
          format: date-time
          type: string
        monthlyRent:
          type: number
        deposit:
          type: number
        canPay:
          type: boolean
        notify:
          type: boolean
        license:
          type: string
          nullable: true
    PortfolioDto:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        fullName:
          type: string
        label:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        civilid:
          type: string
          nullable: true
        dob:
          format: date-time
          type: string
          nullable: true
        organizationId:
          type: string
        title:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - fullName
        - label
        - phone
        - civilid
        - dob
        - organizationId
        - title
        - createdAt
        - updatedAt
...
    PropertyBreadcrumbsDto:
      type: object
      properties:
        portfolio:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/BreadcrumbDto'
        property:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/BreadcrumbDto'
      required:
        - portfolio
        - property
    PropertyDto:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        breadcrumbs:
          $ref: '#/components/schemas/PropertyBreadcrumbsDto'
        title:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        organizationId:
          type: string
        portfolioId:
          type: string
        area:
          type: string
          nullable: true
        block:
          type: string
          nullable: true
        avenue:
          type: string
          nullable: true
        street:
          type: string
          nullable: true
        number:
          type: string
          nullable: true
        parcel:
          type: string
          nullable: true
        paci:
          type: string
          nullable: true
        cost:
          type: number
          nullable: true
        label:
          type: string
          nullable: true
      required:
        - id
        - breadcrumbs
        - title
        - createdAt
        - updatedAt
        - organizationId
        - portfolioId
        - area
        - block
        - avenue
        - street
        - number
        - parcel
        - paci
        - cost
        - label
    UnitVacancy:
      type: object
      properties:
        isVacant:
          type: boolean
        vacancyDistance:
          type: string
        vacancyDate:
          format: date-time
          type: string
      required:
        - isVacant
        - vacancyDistance
        - vacancyDate
...
    PayoutDto:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        breadcrumbs:
          $ref: '#/components/schemas/PayoutBreadcrumbsDto'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        organizationId:
          type: string
        portfolioId:
          type: string
        amount:
          type: number
        postAt:
          format: date-time
          type: string
        memo:
          type: string
          nullable: true
      required:
        - id
        - breadcrumbs
        - createdAt
        - updatedAt
        - organizationId
        - portfolioId
        - amount
        - postAt
...
    ExpenseCategoryDto:
      type: object
      properties:
        id:
          type: string
        labelEn:
          type: string
        labelAr:
          type: string
          nullable: true
        parentId:
          type: string
          nullable: true
        isGroup:
          type: boolean
      required:
        - id
        - labelEn
        - labelAr
        - parentId
        - isGroup
...
    PartialUnitDto:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        vacancy:
          $ref: '#/components/schemas/UnitVacancy'
        breadcrumbs:
          $ref: '#/components/schemas/UnitBreadcrumbsDto'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        organizationId:
          type: string
        portfolioId:
          type: string
        propertyId:
          type: string
        unitNumber:
          type: string
        floor:
          type: number
          nullable: true
        size:
          type: number
          nullable: true
        bed:
          type: number
          nullable: true
        bath:
          type: number
          nullable: true
        marketRent:
          type: number
          nullable: true
        type:
          type: string
          nullable: true
        usage:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
...
    PartialExpenseDto:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        expenseType:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ExpenseCategoryDto'
        breadcrumbs:
          $ref: '#/components/schemas/ExpenseBreadcrumbsDto'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        portfolioId:
          type: string
        propertyId:
          type: string
          nullable: true
        unitId:
          type: string
          nullable: true
        amount:
          type: number
        postAt:
          type: string
        categoryId:
          type: string
          nullable: true
        memo:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        organizationId:
          type: string
...
    CreateTenantDto:
      type: object
      properties:
        fullName:
          type: string
        label:
          type: string
          nullable: true
        civilid:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        dob:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        passportNum:
          type: string
          nullable: true
        residencyEnd:
          type: string
          nullable: true
        residencyNum:
          type: string
          nullable: true
      required:
        - fullName
    CreatePortfolioDto:
      type: object
      properties:
        fullName:
          type: string
        label:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        civilid:
          type: string
          nullable: true
        dob:
          type: string
          nullable: true
      required:
        - fullName
    CreatePropertyDto:
      type: object
      properties:
        portfolioId:
          type: string
        number:
          type: string
        area:
          type: string
        block:
          type: string
        street:
          type: string
        label:
          type: string
          nullable: true
        avenue:
          type: string
          nullable: true
        parcel:
          type: string
          nullable: true
        paci:
          type: string
          nullable: true
      required:
        - portfolioId
        - number
        - area
        - block
        - street
    CreateUnitDto:
      type: object
      properties:
        portfolioId:
          type: string
        propertyId:
          type: string
        unitNumber:
          type: string
        marketRent:
          type: number
          nullable: true
        label:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        bed:
          type: number
          nullable: true
        bath:
          type: number
          nullable: true
        size:
          type: number
          nullable: true
        floor:
          type: number
          nullable: true
        usage:
          type: string
          nullable: true
      required:
        - portfolioId
        - propertyId
        - unitNumber
    CreateLeaseDto:
      type: object
      properties:
        portfolioId:
          type: string
        unitId:
          type: string
        tenantId:
          type: string
        start:
          type: string
        monthlyRent:
          type: number
        deposit:
          type: number
        end:
          type: string
        notify:
          type: boolean
        canPay:
          type: boolean
        license:
          type: string
          nullable: true
      required:
        - portfolioId
        - unitId
        - tenantId
        - start
        - monthlyRent
        - deposit
        - end
        - notify
        - canPay
    CreateLeaseInvoiceDto:
      type: object
      properties:
        portfolioId:
          type: string
        leaseId:
          type: string
        amount:
          type: number
        isPaid:
          type: boolean
        postAt:
          type: string
        dueAt:
          type: string
          nullable: true
        paidAt:
          type: string
          nullable: true
        memo:
          type: string
          nullable: true
      required:
        - portfolioId
        - leaseId
        - amount
        - isPaid
        - postAt
    CreateManyLeaseInvoicesDto:
      type: object
      properties:
        portfolioId:
          type: string
        amount:
          type: number
        isPaid:
          type: boolean
        postAt:
          type: string
        dueAt:
          type: string
          nullable: true
        paidAt:
          type: string
          nullable: true
        memo:
          type: string
          nullable: true
      required:
        - portfolioId
        - amount
        - isPaid
        - postAt
    CreateExpenseDto:
      type: object
      properties:
        portfolioId:
          type: string
        propertyId:
          type: string
          nullable: true
        unitId:
          type: string
          nullable: true
        amount:
          type: number
        postAt:
          type: string
        categoryId:
          type: string
          nullable: true
        memo:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
      required:
        - portfolioId
        - propertyId
        - unitId
        - amount
        - postAt
    CreatePayoutDto:
      type: object
      properties:
        portfolioId:
          type: string
        amount:
          type: number
        postAt:
          type: string
        memo:
          type: string
          nullable: true
      required:
        - portfolioId
        - amount
        - postAt
    FileRelationKeyEnum:
      type: string
      enum:
        - tenant
        - portfolio
        - property
        - unit
        - expense
        - lease
        - leaseInvoice
        - maintenanceOrder
    CreateFileDto:
      type: object
      properties:
        relationKey:
          $ref: '#/components/schemas/FileRelationKeyEnum'
        file:
          type: string
          format: binary
        fileName:
          type: string
        relationValue:
          type: string
      required:
        - relationKey
        - file
        - fileName
        - relationValue
    FileDto:
      type: object
      properties:
        id:
          type: string
        size:
          type: number
      required:
        - id
        - size
...
    PaginatedMetaDto:
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/PaginatedDto'
      required:
        - pagination